#276 fix: 피드백 반영

+++ src/main/java/UMC_7th/Closit/domain/battle/dto/BattleDTO/BattleResponseDTO.java
@@ -63,9 +63,9 @@ public class BattleResponseDTO {
     public static class VoteBattleResultDTO { // 배틀 투표
         private Long battleId;
         private String firstClositId;
-        private double firstVotingCnt;
+        private int firstVotingCnt;
         private String secondClositId;
-        private double secondVotingCnt;
+        private int secondVotingCnt;
         @JsonFormat(pattern = "yyyy/MM/dd HH:mm:ss")
         private LocalDateTime createdAt;
     }
@@ -83,13 +83,13 @@ public class BattleResponseDTO {
         private Long firstPostId;
         private String firstPostFrontImage;
         private String firstPostBackImage;
-        private double firstVotingCnt;
+        private int firstVotingCnt;
         private String secondClositId;
         private String secondProfileImage;
         private Long secondPostId;
         private String secondPostFrontImage;
         private String secondPostBackImage;
-        private double secondVotingCnt;
+        private int secondVotingCnt;
     }
 
     @Builder
@@ -145,13 +145,13 @@ public class BattleResponseDTO {
         private Long firstPostId;
         private String firstPostFrontImage;
         private String firstPostBackImage;
-        private double firstVotingCnt;
+        private int firstVotingCnt;
         private String secondClositId;
         private String secondProfileImage;
         private Long secondPostId;
         private String secondPostFrontImage;
         private String secondPostBackImage;
-        private double secondVotingCnt;
+        private int secondVotingCnt;
     }
 
     @Builder
